ebirdcbc-backend/pyproject.toml
[project]
name = "ebirdcbc-v2"
version = "0.1.0"
description = "Backend for eBird CBC App"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "bcrypt>=4.2.1",
    "cryptography>=44.0.0",
    "fastapi[standard]>=0.115.6",
    "playwright>=1.49.1",
    "polars>=1.17.1",
    "pydantic>=2.10.4",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "supabase>=2.10.0",
]
